Lonely release
A doctor, an engineer, and a lawyer are sentenced to death by beheading from a guillotine in France. 

The doctor goes first, but as the blade comes down it stops halfway. This is thought of as a sign from god and the man is set free. 

The lawyer goes next, and again the blade stops halfway and the man is let go.

Finally the engineer goes, but just as the executioner is about to drop the blade he says "Oh I see the problem!"